Fix slight pixel error on responsive view.
The previous code was hacked together to account for Chrome's scrollbar and thus would bounce with Firefox's scrollbar that is not part of the viewport. The correct answer is to use the container query instead of the media query to base the pixel measurements off the view itself, not the viewport.